Use float32 for startup timestamp sync
The startup-time MIN all_reduce used torch.double tensors. ECCL
(Enflame GCU's NCCL substitute) rejects fp64 (and int64) all_reduce
with ecclInvalidArgument, aborting training before the first step.
fp32 is accepted and sufficient here: the values are only compared
with MIN and used for startup bookkeeping, so the coarse precision of
fp32 at epoch-scale timestamps is an acceptable tradeoff for portability
across non-CUDA collectives.

+++ megatron/training/training.py
@@ -1039,13 +1039,13 @@ def pretrain(
     # Initialize program_start_global with a fallback value in case set_startup_timestamps() wasn't called
     program_start_global = _TRAIN_START_TIME
     if _STARTUP_TIMESTAMPS['program_start'] is not None:
-        program_start_global = torch.tensor([_STARTUP_TIMESTAMPS['program_start']], dtype=torch.double, device='cuda')
+        program_start_global = torch.tensor([_STARTUP_TIMESTAMPS['program_start']], dtype=torch.float32, device='cuda')
         torch.distributed.all_reduce(program_start_global, op=torch.distributed.ReduceOp.MIN)
         program_start_global = program_start_global.item()
     set_startup_timestamps(program_start=program_start_global)
 
     global _LEGACY_TRAIN_START_TIME
-    start_time_tensor = torch.tensor([_LEGACY_TRAIN_START_TIME], dtype=torch.double, device='cuda')
+    start_time_tensor = torch.tensor([_LEGACY_TRAIN_START_TIME], dtype=torch.float32, device='cuda')
     torch.distributed.all_reduce(start_time_tensor, op=torch.distributed.ReduceOp.MIN)
     _LEGACY_TRAIN_START_TIME = start_time_tensor.item()
